pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bespokeasm"
version = "0.7.3c2"
authors = [
  { name="Michael Kamprath", email="michael@kamprath.net" },
]
description = "A customizable byte code assembler that allows for the definition of custom instruction set architecture"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Assemblers",
]
dependencies = [
    "click",
    "intelhex",
    "packaging",
    "pre-commit",
    "click-default-group",
    "ruamel.yaml",
]

[project.optional-dependencies]
dev = [
    "build",
    "pipx",
    "pyinstaller",
    "nuitka[onefile]",
]

[project.urls]
homepage = "https://github.com/michaelkamprath/bespokeasm"
repository = "https://github.com/michaelkamprath/bespokeasm"
documentation = "https://github.com/michaelkamprath/bespokeasm/wiki"
changelog = "https://github.com/michaelkamprath/bespokeasm/blob/main/CHANGELOG.md"

[project.scripts]
bespokeasm = "bespokeasm.__main__:entry_point"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.xml", "*.yaml", "*.js"]
